requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
opencv-contrib-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
